Refactored to create object model, added min event size
Apples ical visually overlays events if they don't overlap in time.  This fix extends each event to be at least 8mins
that way events close to each other will overlap in time which causes them to be layed out next to each other instead
of ontop of each other.
To achived this there is now a data model, which represent the visit trails, and then it is turned into ical events.